requirements.txt
Flask
Flask-Limiter
RPi.GPIO
